Script location fix

+++ post_install.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+cd "$(dirname "$0")" || exit
+
 # shellcheck source=./shared/colors.sh
 source "./shared/colors.sh"
 
@@ -13,8 +15,41 @@ resolve_distro() {
   esac
 }
 
+# Check the package manager
+get_package_manager() {
+    if [ -x "/usr/bin/zypper" ]; then
+     echo "zypper"
+  elif [ -x "/usr/bin/dnf" ]; then
+    echo "dnf"
+  elif [ -x "/usr/bin/pacman" ]; then
+    echo "pacman"
+  elif [ -x "/usr/bin/apt" ]; then
+    echo "apt"
+  else
+    echo "unknown"
+  fi
+}
+
+package_manager=$(get_package_manager)
+
+# check if whiptail is installed and install it
+if [ ! -x "/usr/bin/whiptail" ]; then
+  echo -e "${RED}whiptail is not installed! Please install newt to proceed!${NC}"
+
+  case $package_manager in
+    "zypper") sudo zypper -vv in -y newt;;
+    "dnf") sudo dnf in -y newt;;
+    "pacman") sudo pacman -S --noconfirm whiptail;;
+    "apt") sudo apt install -y whiptail;;
+    *) echo -e "${RED}Couldn't detect package manager!${NC}"
+       exit 1;;
+  esac
+
+fi
+
 whiptail --title "Linux Post-Install Script" --msgbox "Welcome to the post install script!\nFirst we'll need to gather some info about your system" 0 0
 
+# Auto detect distro
 if grep -iq opensuse /etc/os-release; then
   chosen_distro="1"
 elif grep -iq fedora /etc/os-release; then
@@ -23,6 +58,8 @@ elif grep -iq "arch linux" /etc/os-release; then
   chosen_distro="3"
 elif grep -iq debian /etc/os-release; then
   chosen_distro="4"
+else
+  chosen_distro="-1"
 fi
 
 echo -e "${GREEN}$(resolve_distro "$chosen_distro") detected!${NC}"
@@ -42,19 +79,23 @@ fi
 echo -e "${GREEN}Your chosen distro is $(resolve_distro "$chosen_distro")${NC}"
 
 echo -e "${GREEN}Checking package manager...${NC}"
-if [[ $chosen_distro = "1" && -x "/usr/bin/zypper" ]]; then
+if [[ $chosen_distro = "1" && $package_manager = "zypper" ]]; then
   echo -e "${GREEN}zypper found for OpenSUSE!${NC}"
   sh "./distros/opensuse/setup.sh"
-elif [[ $chosen_distro = "2" && -x "/usr/bin/dnf" ]]; then
+elif [[ $chosen_distro = "2" && $package_manager = "dnf" ]]; then
   echo -e "${GREEN}dnf found for Fedora!${NC}"
   sh "./distros/fedora/setup.sh"
-elif [[ $chosen_distro = "3" && -x "/usr/bin/pacman" ]]; then
+elif [[ $chosen_distro = "3" && $package_manager = "pacman" ]]; then
   echo -e "${GREEN}pacman found for Arch linux!${NC}"
   sh "./distros/arch/setup.sh"
-elif [[ $chosen_distro = "4" && -x "/usr/bin/apt" ]]; then
+elif [[ $chosen_distro = "4" && $package_manager = "apt" ]]; then
   echo -e "${GREEN}apt found for Debian!${NC}"
   sh "./distros/debian/setup.sh"
 else
   echo -e "${RED}Can't continue! Mismatched package manager and distro!${NC}"
   exit 1
 fi
+
+echo -e "${YELLOW}Please reboot for flatpak's path and QEMU to work${NC}"
+echo -e "${YELLOW}Please run 'gh auth login' to start using GitHub CLI${NC}"
+echo -e "${GREEN}Post install complete, enjoy your new distro!${NC}"
